Description
In these files there are two  CSV files . Through some basic calculation and using the power bi tool , we are  able to  create a beautiful dashboard . It will help us to know the business better and also helps us to improve the business . 
There is a pdf file which is the final dashboard created by me. 
There is a power bi file also . If you want , you can try it out.